requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-dotenv>=1.0.1

# OAuth and HTTP
aiohttp>=3.9.3
requests-oauthlib>=1.3.1
tweepy>=4.14.0

# Database and caching
SQLAlchemy>=2.0.27
aiosqlite>=0.19.0
redis>=5.0.1

# Security
cryptography>=42.0.5
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9
